+++ packages/vendure-plugin-address-lookup/src/constants.ts
@@ -1,0 +1,4 @@
+export const ADDRESS_LOOKUP_PLUGIN_OPTIONS = Symbol(
+  'ADDRESS_LOOKUP_PLUGIN_OPTIONS'
+);
+export const loggerCtx = 'AddressLookupPlugin';