+++ packages/vendure-plugin-address-lookup/src/index.ts
@@ -1,0 +1,5 @@
+export * from './address-lookup.plugin';
+export * from './types';
+export * from './generated/graphql';
+export * from './config/post-nl-lookup-strategy';
+export * from './config/postcode-tech-strategy';